Extract shared icon.sys helpers and add UI tests
